+++ docs/source/conf.py
@@ -9,7 +9,7 @@
 project = 'demoutrei.discord'
 copyright = '2026, demoutrei'
 author = 'demoutrei'
-release = '26.1.11'
+release = '26.1.12-dev'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -19,10 +19,6 @@ extensions = ["sphinx.ext.autodoc", "sphinx.ext.viewcode", "sphinx.ext.intersphi
 autodoc_default_options = {
   "members": True,
   "undoc-members": True
-}
-
-intersphinx_mapping = {
-  "python": ("https://docs.python.org/3", None)
 }
 
 toc_object_entries_show_parents = "hide"